ImageCollector test renamed
