+++ Tests/Samples/Samples.cs
@@ -8,6 +8,7 @@ namespace Test
 		public const string IMAGESDIR = "../../../Samples/Images";
 		public const string CBZPATH = "../../../Samples/Crossed 3D.cbz";
 		public const string CBRPATH = "../../../Samples/CROSSED Wish you were here.cbr";
-		public const string FAKEFILEPATH = "../../../Samples/File.test";
+		public const string TESTPATH = "../../../Samples/File.test";
+		public const string FAKEFILE  = "Do not exist.txt";
 	}
 }
